django_microservice/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
